# DomainIntelWeb/openapi.json
{
  "components": {
    "schemas": {
      "ArchiveState": {
        "properties": {
          "archived_to": {
            "title": "Archived To",
            "type": "string"
          }
        },
        "required": [
          "archived_to"
        ],
        "title": "ArchiveState",
        "type": "object"
      },
...
      "CoverageState": {
        "properties": {
          "cells": {
            "items": {
              "$ref": "#/components/schemas/CoverageCellState"
            },
            "title": "Cells",
            "type": "array"
          },
          "summary": {
            "$ref": "#/components/schemas/CoverageSummaryState"
          }
        },
        "required": [
          "cells",
          "summary"
        ],
        "title": "CoverageState",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HistoryCoverageState": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/HistoryHorizonState"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "HistoryCoverageState",
        "type": "object"
      },
...
      "IndustryMutationState": {
        "properties": {
          "folder": {
            "title": "Folder",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "folder",
          "name"
        ],
        "title": "IndustryMutationState",
        "type": "object"
      },
      "IndustryRename": {
        "properties": {
          "folder": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Folder",
            "type": "string"
          },
          "name": {
            "default": "",
            "maxLength": 120,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "folder"
        ],
        "title": "IndustryRename",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "IntDog Local API",
    "version": "4.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/industries/{folder}": {
      "delete": {
        "operationId": "archive_industry_api_industries__folder__delete",
        "parameters": [
          {
            "in": "path",
            "name": "folder",
            "required": true,
            "schema": {
              "title": "Folder",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArchiveState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Archive Industry",
        "tags": [
          "industries"
        ]
      },
      "patch": {
        "operationId": "rename_industry_api_industries__folder__patch",
        "parameters": [
          {
            "in": "path",
            "name": "folder",
            "required": true,
            "schema": {
              "title": "Folder",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IndustryRename"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndustryMutationState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rename Industry",
        "tags": [
          "industries"
        ]
      }
    },
...
    "/api/industries/{folder}/coverage": {
      "get": {
        "operationId": "coverage_api_industries__folder__coverage_get",
        "parameters": [
          {
            "in": "path",
            "name": "folder",
            "required": true,
            "schema": {
              "title": "Folder",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoverageState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Coverage",
        "tags": [
          "intelligence"
        ]
      }
    },
    "/api/industries/{folder}/coverage/cells": {
      "post": {
        "operationId": "create_coverage_cell_api_industries__folder__coverage_cells_post",
        "parameters": [
          {
            "in": "path",
            "name": "folder",
            "required": true,
            "schema": {
              "title": "Folder",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CoverageCellCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoverageCellState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Coverage Cell",
        "tags": [
          "intelligence"
        ]
      }
    },
    "/api/industries/{folder}/coverage/cells/{cell_id}/attempts": {
      "post": {
        "operationId": "record_attempt_api_industries__folder__coverage_cells__cell_id__attempts_post",
        "parameters": [
          {
            "in": "path",
            "name": "folder",
            "required": true,
            "schema": {
              "title": "Folder",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "cell_id",
            "required": true,
            "schema": {
              "title": "Cell Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CoverageAttemptCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoverageAttemptState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record Attempt",
        "tags": [
          "intelligence"
        ]
      }
    },
    "/api/industries/{folder}/coverage/initialize": {
      "post": {
        "operationId": "initialize_coverage_api_industries__folder__coverage_initialize_post",
        "parameters": [
          {
            "in": "path",
            "name": "folder",
            "required": true,
            "schema": {
              "title": "Folder",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Initialize Coverage Api Industries  Folder  Coverage Initialize Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Initialize Coverage",
        "tags": [
          "intelligence"
        ]
      }
    },
    "/api/industries/{folder}/coverage/plan": {
      "post": {
        "operationId": "plan_coverage_api_industries__folder__coverage_plan_post",
        "parameters": [
          {
            "in": "path",
            "name": "folder",
            "required": true,
            "schema": {
              "title": "Folder",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Plan Coverage Api Industries  Folder  Coverage Plan Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Plan Coverage",
        "tags": [
          "intelligence"
        ]
      }
    },
...
    "/api/industries/{folder}/history": {
      "get": {
        "description": "Read live long-horizon coverage without mutating compatibility files.",
        "operationId": "history_api_industries__folder__history_get",
        "parameters": [
          {
            "in": "path",
            "name": "folder",
            "required": true,
            "schema": {
              "title": "Folder",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryCoverageState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "History",
        "tags": [
          "intelligence"
        ]
      }
    },
    "/api/industries/{folder}/knowledge/entities": {
      "get": {
        "operationId": "knowledge_entities_api_industries__folder__knowledge_entities_get",
        "parameters": [
          {
            "in": "path",
            "name": "folder",
            "required": true,
            "schema": {
              "title": "Folder",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "query",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 200,
              "title": "Query",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 80,
              "title": "Kind",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "country",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 100,
              "title": "Country",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 80,
              "title": "Status",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "chain",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 200,
              "title": "Chain",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeEntityPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Knowledge Entities",
        "tags": [
          "intelligence"
        ]
      }
    },
    "/api/industries/{folder}/knowledge/entities/{entity_id}": {
      "get": {
        "operationId": "knowledge_entity_api_industries__folder__knowledge_entities__entity_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "folder",
            "required": true,
            "schema": {
              "title": "Folder",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "entity_id",
            "required": true,
            "schema": {
              "title": "Entity Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeEntityDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Knowledge Entity",
        "tags": [
          "intelligence"
        ]
      }
    },
...
    "/api/industries/{folder}/stories": {
      "get": {
        "operationId": "stories_api_industries__folder__stories_get",
        "parameters": [
          {
            "in": "path",
            "name": "folder",
            "required": true,
            "schema": {
              "title": "Folder",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoryListState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stories",
        "tags": [
          "intelligence"
        ]
      }
    },
    "/api/industries/{folder}/stories/{story_id}": {
      "get": {
        "operationId": "story_api_industries__folder__stories__story_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "folder",
            "required": true,
            "schema": {
              "title": "Folder",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "story_id",
            "required": true,
            "schema": {
              "title": "Story Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoryDetailState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Story",
        "tags": [
          "intelligence"
        ]
      }
    },
    "/api/industries/{folder}/stories/{story_id}/merge": {
      "post": {
        "operationId": "merge_story_api_industries__folder__stories__story_id__merge_post",
        "parameters": [
          {
            "in": "path",
            "name": "folder",
            "required": true,
            "schema": {
              "title": "Folder",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "story_id",
            "required": true,
            "schema": {
              "title": "Story Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoryMergeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoryDetailState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Merge Story",
        "tags": [
          "intelligence"
        ]
      }
    },
    "/api/industries/{folder}/stories/{story_id}/split": {
      "post": {
        "operationId": "split_story_api_industries__folder__stories__story_id__split_post",
        "parameters": [
          {
            "in": "path",
            "name": "folder",
            "required": true,
            "schema": {
              "title": "Folder",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "story_id",
            "required": true,
            "schema": {
              "title": "Story Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StorySplitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoryDetailState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Split Story",
        "tags": [
          "intelligence"
        ]
      }
    },
    "/api/industries/{folder}/stories/{story_id}/unlock": {
      "post": {
        "operationId": "unlock_story_api_industries__folder__stories__story_id__unlock_post",
        "parameters": [
          {
            "in": "path",
            "name": "folder",
            "required": true,
            "schema": {
              "title": "Folder",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "story_id",
            "required": true,
            "schema": {
              "title": "Story Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoryUnlockRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoryDetailState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Unlock Story",
        "tags": [
          "intelligence"
        ]
      }
    },
...
  }
}